IpcSocketFactory pending packet limit is too small
Refactoring is underway to notify client of P2P socket that network is
congested. In the mean time pending packet limit is too small and causes
huge amount of packet loss, this makes WebRTC unusable.

This is a temporary fix to increase this pre-defined value to 32 pending
packets which accomodates 720p streams up to 5Mb/s.

BUG=228956, 229737

Review URL: https://codereview.chromium.org/13992004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193389 0039d316-1c4b-4281-b951-d872f2087c98