nacl-sdk: Add a PNaCl finalize step to the Release builds for SDK examples.
For PNaCl bitcode stability we need to strip metadata, such as debug
since the debug metadata is still evolving.  We want to show developers
that this is part of the workflow.

Add a release build of the examples to test this out.
Adds ~1 min 40 to windows build.


BUG= https://code.google.com/p/nativeclient/issues/detail?id=3348
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/13466021

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193671 0039d316-1c4b-4281-b951-d872f2087c98