Remaining issues with making the autofill dialog usable
This fixes the initialization so that updating menu items dont trigger
layout changes. It makes the dialog resize on keyboard coming up and down.
It also adds an email layout so that we have a valid(although non-ideal)
transition for all sections. The transition will be fixed properly on
crbug.com/180980

BUG=226500,226496,226497
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/14165005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193732 0039d316-1c4b-4281-b951-d872f2087c98