Use the windows path for reposting menu events on win aura
On non-Aura, the behaviour is that when the menu is open, presses
outside of the menu (say, to start a tab drag, or to click "new tab")
both dismiss the menu and perform the mouse press action.

This changes win aura to match win non-aura.

BUG=229983

Review URL: https://chromiumcodereview.appspot.com/14029019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193804 0039d316-1c4b-4281-b951-d872f2087c98