Remove some uninitialized reads from simple cache metadata serializer.
Valgrind detects uninitialized reads very well, and that's great! It catches cases when we allocate garbage on stack, hash it, write to the file. If we ever had wanted to repeat the operation, the hash would have been most likely be different, which could be fine in prod, but bad for testing.

More details in the bug.

BUG=173393

Review URL: https://codereview.chromium.org/14204002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@193936 0039d316-1c4b-4281-b951-d872f2087c98