Instant: Don't send autocomplete results if the user isn't typing.
InstantController::Update() sends onchange(query) to the page only if the user
is actively typing (more specifically, if the popup is open). Likewise, we
should only send the autocomplete results if the popup is open.

This avoids the page receiving spurious autocomplete-results-changed events,
especially when the popup is being closed (due to the user clicking away).

BUG=225326
R=pkasting@chromium.org
TEST=See bug (the first case in comment #0; and the case in comment #5).

Review URL: https://chromiumcodereview.appspot.com/13839006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@194009 0039d316-1c4b-4281-b951-d872f2087c98