We need to call aura::Window::Show from the DesktopNativeWidgetAura::ShowMaximizedWithBounds and ShowWithWindowState functions.
This is specifically important if the underlying Widget was hidden and then made visible via Widget::Show. The visibility state of
the aura Window object remains as hidden, which results in the compositor not painting this window.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=229913 where the composited panel window would turn black on Windows 8 if we
switched to the metro screen and back.

R=ben@chromium.org
BUG=229913
TEST=Covered by new views unit test TestWindowVisibilityAfterHide

Review URL: https://chromiumcodereview.appspot.com/13849012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@195270 0039d316-1c4b-4281-b951-d872f2087c98