Fixes a crash caused for update of webkit notifications.
If update happens with different ids, it essentially means a removal
and an add from notification_list point of view. Sending Update() will
confuse the observers.

It is possible to pass both ids for Update(), but handling update of
different ids on each observer doesn't make much sense. It will be better
to send both Removed() and Added() in this case.

BUG=233221
TEST=manually on the reported page in the bug

Review URL: https://chromiumcodereview.appspot.com/14127003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@195346 0039d316-1c4b-4281-b951-d872f2087c98