Fixed a bug that was introduced when we started deprecating 32bit libraries on 64bit systems.
This needs a little bit of explanation. Normally, we would want to restructure
the code to have separate variables for "--lib32" and for the result of the
"yes_no" question. But that requires more global changes.

Instead, in this case, we want to very clearly document that we added an
early "if" statement to break out. And this is intended to be a temporary
measure until the entire feature is fully removed.

So, I instead opted for resetting the "do_inst_lib32" variable -- and a lengthy
comment explaining this decision.

BUG=233047
TEST=run with --lib32 and verify that entering "N" aborts the installation.

Review URL: https://chromiumcodereview.appspot.com/14166013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@195950 0039d316-1c4b-4281-b951-d872f2087c98