Launching ASH on Windows 8 should activate an existing instance if one exits.
The current behavior when there is an existing instance of ASH is that the launch fails taking
down the existing instance of ASH.

This is because metro app launches go through the delegate execute handler which first tries to launch the
metro viewer process (browser). If we have an existing browser process which has an active MetroViewerProcessHost
instance, we delete it first and recreate it. This takes down the existing ASH instance and attempts to launch the
new one fail, because of a race condition in the metro startup code.

Fix is to create a new instance of the MetroViewerProcessHost instance only if one does not exist in the 
BrowserProcessImpl::PerformInitForWindowsAura function which is invoked when a brower process processes
a launch command line.

We reset the MetroViewerProcessHost instance when we receive an OnChannelError notification in the MetroViewerProcessHost 
class. It in turn invokes the virtual function OnMetroViewerProcessTerminated added on the BrowserProcess interface with
a windows aura specific implementation in the browser_process_impl_win.cc file.

BUG=234429
R=cpu

Review URL: https://codereview.chromium.org/13973023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@196254 0039d316-1c4b-4281-b951-d872f2087c98