Aura: prevent race in resize lock handling causing content to stop updating.
Because the renderer may not see all the size changes (from RWH throttling),
keeping a list of locks for all previous size changes makes it impossible to
figure out which resize an incoming frame is for. It's usually not a problem,
except when the UI is:
1- ignoring the "hold mouse moves" logic
2- flipping many times between the same 2 sizes faster than the renderer can
   handle.
In that case we may end up in a situation where we still have locks in place,
but the renderer is done with size changes, so we'll never pop the remaining
locks, and we will skip every renderer frame (ShouldSkipFrame returns true).

Instead, change the logic so that we keep a single ResizeLock, until we receive
and commit the frame of the right size. We don't send more resizes to the
renderer until then.

BUG=233717

Review URL: https://chromiumcodereview.appspot.com/14451002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@196305 0039d316-1c4b-4281-b951-d872f2087c98