Reland "Swap touch HUD when switching primary display"
When a display is set as the primary display, its root window (and all
its children, including the widget showing the touch HUD) is replaced
with the old one's. But, the touch HUD widget is bound to display and
should not be swapped between the two displays. So, after swapping the
root windows, touch HUDs are swapped back to their own display.

Original Review URL: https://codereview.chromium.org/13460017/ 

BUG=172237
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14079010

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@196312 0039d316-1c4b-4281-b951-d872f2087c98