Implement Pepper GetDeviceID for Windows
The device ID is generated by mixing a salt which is specific to the users profile and with machine-specific information. This CL also connects the action of clearing the salt with the "Clear browsing data > Deauthrorize content licenses" action.

The ChromeOS implementation of GetDeviceID is unified with the Windows implementation and the old method of generating the IDs has been deprecated.

BUG=235193

Review URL: https://chromiumcodereview.appspot.com/12701002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@196522 0039d316-1c4b-4281-b951-d872f2087c98