gpu: Cleanup 'black-screen on Huawei' work-around.
We got a response from their driver team. The surface
is only corrupted if an FBO is bound when it is first
made-current. So we can avoid re-creating the surface
and just unbind the Fbo instead.

The work-around is still needed in the context itself
for two reasons:
- Virtual context indirections
- The first make-current for new surfaces
  is not in the decoder.

BUG=235935
NOTRY=true

No try, since this doesn't run on mac, so mac_asan bot isn't relevant.

Review URL: https://chromiumcodereview.appspot.com/14021014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@196915 0039d316-1c4b-4281-b951-d872f2087c98