Flip the cc::InputHandler and cc::InputHandlerClient names, they are backwards
The names for the InputHandler and InputHandlerClient are backwards and super
confusing. The InputHandler is the thing that does interesting things with inputs,
like scroll and fling and pinch and such.  The Client is the thing that receives
control messages like animation and fling over updates.

 (╯°□°）╯︵ ┻━┻)

Review URL: https://chromiumcodereview.appspot.com/14021024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@197068 0039d316-1c4b-4281-b951-d872f2087c98