Bluetooth: fix fake device classes
Fake devices are used not only in test cases, but also when running a
Chrome OS build of Chromium on Linux. Fix the inverted keyboard/mouse
classes so that the "Microsoft Mouse" behaves as the real device and
returns false for IsPairable(). This means it can be used to exercise
the Connect-without-Pairing case.

Add a new "Unconnectable Device" for the test that was testing a
device actually failing to connect after being paired.

BUG=none
TEST=device_unittests,
  -Dchromeos=1, chrome --enable-experimental-bluetooth, pair with MS Mouse, verify no pairing dialog appears

R=satorux@chromium.org, youngki@chromium.org

Review URL: https://chromiumcodereview.appspot.com/14273033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@197215 0039d316-1c4b-4281-b951-d872f2087c98