Fixed drag and drop into and out of Browser Plugin.
Only works when the --enable-browser-plugin-compositing flag is used, which causes web_contents_view_guest to be used.

How the flow works:
When a drag event is initiated on the web page, the BrowserPluginGuest receives an DragHostMsg_StartDragging IPC, and keeps track of that fact that it has initiated a drag event, but doesn't handle the IPC.
The IPC is then passed down the chain (not sure by who) and eventually calls WebContentsViewGuest::StartDragging.
WebContentsViewGuest::StartDragging calls the platform dependent WebContentsView of the embedder and starts the OS drag event.

Now, we get updates on the drag event from 2 places, BrowserPlugin::handleDragStatusUpdate, and RenderViewImpl::OnDragTargetDragOver of the embedder process.

BrowserPlugin::handleDragStatusUpdate goes through the guest to call RenderViewHostImpl::DragTargetDragOver on the guest process (but only when the mouse is hovering over the guest)
RenderViewImpl::OnDragTargetDragOver also calls RenderViewHostImpl::DragTargetDragOver but on the embedder process, and it always calls the function, whether or not the mouse is over the guest. For that reason, we keep track of if the mouse if over the guest, and use the embedder to capture and ignore the DragTargetDragOver message on the embedder process when the mouse is over the guest (or we get a flickering mouse).

Now, when the mouse is released, the native drag_host calls RenderViewImpl::OnDragSourceSystemDragEnded, which in turn calls webview()->dragSourceSystemDragEnded(); on the embedder process. However, the guest process webview() is not informed that the drag has ended, so it still blocks user input. So we send a new IPC, DragHostMsg_DragStopped to the embedder to let it inform the guest, who in turn informs the guest webview, that the drag has ended.

BUG= 161112

Review URL: https://chromiumcodereview.appspot.com/12086095

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@198119 0039d316-1c4b-4281-b951-d872f2087c98