Add more information to sync protocol GU messages
This change allows the client to specify a list of 'hints' when sending
a GetUpdate request to the server.  It also adds a set of flags that can
be used to help the server determine whether or not the list of hints is
complete.

This change also adds flags to help the server infer the reason why the
client is requesting updates.  These flags can be set on a per-data type
basis to indicate if there was a local refresh request or a local change
to the type.

BUG=138613,231693

Review URL: https://chromiumcodereview.appspot.com/13831012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@198250 0039d316-1c4b-4281-b951-d872f2087c98