Notify all visible renderers when the visibility of the mouse cursor changes
Blink side patch:
https://codereview.chromium.org/14047016/

The ultimate goal of both patches is to disallow new hover
effects from being invoked in web contents when the mouse
cursor is not visible to the user (i.e., while touch
scrolling). The job of this patch is to communicate the
cursor visibility state to all visible renderers using
the new IPC InputMsg_CursorVisibilityChange whenever the
visibility state changes.

Added the new observer type CursorClientObserver.
Subscribers (instances of RenderWidgetHostViewAura) are
notified whenever the cursor visibility changes, at which
point the IPC is sent to the renderer.

I have also removed the code in ash_native_cursor_manager.cc
that sets the mouse cursor location to the bogus value
of (-10000,-10000) when mouse events are disabled; afaik
this was originally added as a way to prevent unwanted
hover effects in web contents but it does not work in all
cases and causes other problems (see crbug.com/174358).

BUG=153784,174358
R=jamesr@chromium.org, kenrb@chromium.org, oshima@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/14047015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@198519 0039d316-1c4b-4281-b951-d872f2087c98