Add MediaPlayerAndroid as a super class for all media player implementations on android.
There are several different implementations of media players on Android.
Currently they all inherit from MediaPlayerBridge.
This only works if their java player is a subclass of Android MediaPlayer class, which is not necessary in some cases.
Additionally, we are not going to use any subclass of android MediaPlayer for MSE implementation
As a result, this change introduces a super class for all the media player implementations.

BUG=233420
R=scherkus@chromium.org, ycheo@chromium.org, yfriedman@chromium.org

Review URL: https://codereview.chromium.org/14762006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@198957 0039d316-1c4b-4281-b951-d872f2087c98