Removes the space for notification icon if it's not visible.
GridLayout will make the space for a child even if the child is
not visible, which makes an extra space between status area and
panels when there are no notifications.

To remove this space, this CL makes StatusAreaWidgetDelegate catch
child's visibility change, and do not count invisible children in
its GridLayout.

BUG=238770

Review URL: https://chromiumcodereview.appspot.com/14688018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199077 0039d316-1c4b-4281-b951-d872f2087c98