[Memory Sheriff] Fix a leak of MockAsyncResolverBase.
It seems like we were leaking the proxy resolver created in
URLRequestFtpJobTest.FtpProxyRequestOrphanJob.

The original call to set_proxy_service went directly to the request context,
and bypassed the context storage. And URLRequestContext does not take ownership
of the pointers passed in, so that would leak.

After this patch, context_storage_ takes ownership of the proxy service and
cleans up.

TBR=thestig@chromium.org,phajdan@chromium.org
R=mmenke@chromium.org

BUG=238339

Review URL: https://chromiumcodereview.appspot.com/14686008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@199177 0039d316-1c4b-4281-b951-d872f2087c98